requirements.txt
bertopic==0.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.2
hdbscan==0.8.40
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
llvmlite==0.41.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.34.1
networkx==3.2.1
numba==0.58.1
numpy==1.23.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pynndescent==0.5.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.2.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.1.0
tqdm==4.67.1
transformers==4.35.2
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.3
urllib3==2.4.

bertopic
PyPDF2
docx 
python-magic-bin